feat: adds data_mile1 based on Li et al.
Data types include
* landings
* trend
* age composition
* weight-at-age

Uses data-raw to store information on how data_mile1
can be recreated. Adds data-raw to ignore for package build.
Lines with 80 # are used to separate out major sections of the code.
Uses a development branch of the {ASSAMC}
to run a single iteration of the Operating Model but doesn't require
it to be in depends or suggests.