feat: adds weight-at-age data to input data frame
As well as adding weight-at-age data to the input data frame,
this commit also uses a development branch of the {ASSAMC}
to run a single iteration of the Operating Model, and thus,
simplifies the code needed here by quite a bit.

Lines with 80 # are used to separate out major sections of the code.