Updates from 102-feat-data (#176)
* refactor: Uses master of ASSAMC with merge of PR

* doc(data_mile1): Use programatic `r ...`

Calculate the number of rows and columns
using `r ...` rather than hard coding it.

* refactor(FIMSFrame): Standardizes class functions

Use more {methods} functions to validate, set, etc.
the FIMSFrame class. Adds documentation to the fimsframe.R
file to help readers understand what the file contains and
how to potentially augment the file in the future for new
classes. Adds plotting methods for a FIMSFrame.

Removes the FIMSFrame-class documentation because the class
is not actually exported and should not be used. Instead,
the constructors are the only functions that are exported
and documented.