dev: R fxns to setup and run gtest
This initializes functions to set up and run the gtest suite from R. In addition, changes are made to the run-googletest yml to setup for testing. Note that we may want to consider adding "cleanup calls" because files are written to FIMS that are not deleted after running the tests.