feat(FIMSFrame): added start_year, end_year
start_year and end_year are slots in FIMSFrame that are automatically
determined by the data, users do not need to add anything. If they want
to change the start or end year, then they could add dummy data such as
a year of catch for a fleet with zero landings.