feat(create_fims_output):
Make R output constructor

One of three objects available in the FIMS data frame were available to
create_fims_output() via a parameter. Instead, of adding more parameters
I chose to pass the whole data object so accessors could be used to get
the needed dimensionality information. I cheated by creating the data
object from data_mile1 rather than the actual simulated data in om_object
because they have the same dimensions.

The class of items in each slot for FIMSOutput had to be defined
differently. The package_version is not a list it is an S3 class called
package_number and package_version. The latter is now exported. For the
call slot, there currently is no call so I made it NULL instead of NA
and combined the NULL and call classes.

n_fleets is the  number of fleets where fleets is the vector of names
that no longer have to be prefixed with fleet or survey because we do
not care if they are a fleet or a survey.

In the future both fleets and surveys might need to be prefixed with
fleet but right now the fleet/survey distinction is not needed b/c
we do not have any other kinds of data.

Part of #535

TODO: Need to add utils to the import list.