fix(data_mile1): sd of log(x) instead of CV
For catch and index data the measurement of uncertainty should be the
standard deviation (sd) of the logged observations. The operating model
outputs a CV. I added a temporary function in data_mile1.R to calculate
the sd from the CV. In the future we will need to, behind the scenes,
take the log of the sd so the parameters stay within the bounds, which
is what the vignette is doing now. This time series of uncertainty can
later be fed to the modules but I did not change the vignette.

Close #136