fix(create_fims_output): Provide data object
One of three objects available in the FIMS data frame were available to
create_fims_output() via a parameter. Instead, of adding more parameters
I chose to pass the whole data object so accessors could be used to get
the needed dimensionality information. I cheated by creating the data
object from data_mile1 rather than the actual simulated data in om_object
because they have the same dimensions.

nyrs was being called as a function rather than as an object and I
believe this was causing the segfault. Along with similar accessors
also being called incorrectly.

The parameter argument was incorrect for sdr and the documentation was
a different parameter name. I also specified utils:: and added it to
the import list.

Lastly, the class of items in each slot for FIMSOutput had to be defined
differently. The package_version is not a list it is an S3 class called
package_number and package_version. The latter is now exported. For the
call slot, there currently is no call so I made it NULL instead of NA
and combined the NULL and call classes.