refactor: Deletes issue and PR templates
They have been moved and updated in NOAA-FIMS/.github/.github
with NOAA-FIMS/.github@33c261d. Now they are available to all repositories
not just NOAA-FIMS/FIMS. I did not migrate over the refactor_request
issue template because it was largely redundant with the developer_other_issue
template and I thought it was not needed.