feat(FIMSFrame): Add n_fleets
n_fleets is the  number of fleets where fleets is the vector of names
that no longer have to be prefixed with fleet or survey because we do
not care if they are a fleet or a survey.

In the future both fleets and surveys might need to be prefixed with
fleet but right now the fleet/survey distinction is not needed b/c
we do not have any other kinds of data.

Part of #535