fix(FIMSOutput): Get all info out of sdreport
Use all the information in sdreport rather than relying on specific
functions to pull out the values. Indexing vectors are created and then
they are used based on the current order of variables. This could be
made smarter in the future. There are a lot of notes in the .R file on
things that need changed or augmented for the future.

Committing now and giving Sam output.