fix(proportion_female): Remove user interface
Users thought that they could set proportion_female but it was hard-
coded to 0.5 on the back end. Now, it is back to just being a single
value that is hard coded to 0.5. Tests have also changed back to 0.5.
This partially reverts PR #543. And, was talked about in
NOAA-FIMS/seaside-chats#7 and partially addresses #638.