doc(rcpp): Uses 3rd entry to document rcpp
Was PR #652 and commit e0fbfb091ff27ac365481fae79396cbdf036ad02 but
that was based on main and enough has changed that the commit had be
redone for dev. There are still some interfaces that are not documented
but they can be completed in #662.

Thanks to @Bai-Li-NOAA for this commit :)

Close #635