fix(plot.FIMSFrame): Adds y and ...
For plot to be properly documented and rendered all arguments must be
present in the methods, previously we only had x, which led to scales
not being seen as a package that was needed. Now that y and ... have
been added the package is compiling without the error noted by

Part of #671
Close the PR #672, which is no longer needed with this commit. The
methods proposed in #672 was a work around where this commit fixes the
problem.