feat(allcontributors): Adds avatars to README
Uses {allcontributors} in a GitHub action to add 3 lists of contributors
to the README. Also, updates the DESCRIPTION file to reflect Kelli
being the project lead, authors have committed code, and contributors
have contributed code at some time but are not active authors. Some of
these decisions were discussed during a seaside chat on August 22, 2024
and some was discussed during an Implementation Team Meeting on
October 23, 2024. We will be using our own GitHub action unless the
action can be moved to GitHub actions for R and created into a reusable
workflow.