refactor(D*Distribution): Removes TMB from names
Removes TMB from Distribution names and files to clean up the code and
allow the downstream R interface to also remove these names. This is part
of ensuring that the code is portable. If we change from TMB in the future
we do not want to have to rename these things when a distribution is
just a distribution, it does not have to come from TMB. Thank you
@msupernaw for these changes.