chore: Adds documentation to S4 methods
m_weight_at_age and Rcpp_ParameterVector needed help on the
documentation, where none of the methods were documented for the latter.
They still are not exported, just the m_* functions are.

Moves Rcpp_Parameter* to zzz.R because when in operators it was leading to
warnings when compiling in R. The movement allows this file to be ran before
the others and it does not matter if Rcpp_Parameter or Rcpp_ParameterVector is
exported/defined. I also worked on the formatting of those functions as well as
combining the documentation rather for similar, i.e., methods.

Part of #671