refactor(get_*, m_*) feat(get_*): Accessors
Refactor, increase, and overhaul accessors for FIMSFrame and FIMSFit!
Changes to use get_*() and m_*() for all accessors where it was not
consistent before. Also changed to using a single file to document all
of the accessors that can be grouped together.

Added a ton of documentation, i.e., R comments, to the fimsframe and
fimsfit files so it is more clear how to add to them if you are a new
developer.

Increased the consistency of naming and documentation styles, added
VS Code/RStudio style headers to the two files to help with navigation.

Deleted weight_at_age slot from FIMSFrame because m_weight_at_age should
be used instead of getting that information from the slot.

HUGE THANKS to those that showed up and helped with this large task
during the FIMS Office Hour today :)

Co-authored-by: @JaneSullivan-NOAA <jane.sullivan@noaa.gov>
Co-authored-by: @Andrea-Havron-NOAA <andrea.havron@noaa.gov>
Co-authored-by: @nathanvaughan-NOAA <nathan.vaughan@noaa.gov>
Co-authored-by: @peterkuriyama-NOAA <peter.kuriyama@noaa.gov>