fix(initialize_modules): Returns named list
A non-standard pipe method was being used in the vignettes to be able
to pass the returned output from initialize_modules() to fit_fims(),
where now you can just use a pipe from one function to the next.

Also, fixes my blunder of adding eval = FALSE to two sections of the
fims-demo vignette in an earlier commit.

Fixes the documentation of what is being passed to fit_fims(input = ),
where it is the output of initialize_modules().