fix(allcontributors): alphabetical = TRUE
Now that the fix to allcontributors is in. Thanks @iantaylor-NOAA

Changes commit message to "chore: Updates contributors" to add a descriptor and make the tense present.

PR on dev rather than main, which main is the default and nothing was declared. But we don't want any PR to main.

Updates to use v4 of actions and v7 of create-pull-request