refactor(rcpp): Refactor rcpp interface
Keep all members in a shared pointer allowing copies to be made when an
object goes out of scope. These changes are needed to allow finalize()
to work and not crash the R session. RealVector was created as a part
of this work. ParameterVector are now set using resize() and purrr::walk()
rather than new().

TODO:
* Some additional thought shouldd go into using fromRVector and toRVector
  in future versions of this code. See #766
* Update zzz.R to set methods for RealVector that handle indexing
  translation (e.g., R uses 1-based indexing, C++ uses 0-based indexing).
* Modify the relevant R code that assigns values to RealVector to
  accommodate the updated indexing.