Dev: Fleet vs Index comps
Added switch to comps to sample from landings when landings are reported and sample from index when no landings are present. Comps are different for each because survey uses the whole population and landings are just of the retained portion which is not exactly the same due to the impact of natural mortality.