feat(CPUE): Allows for CPUE data and landings
Allows catch-per-unit effort data with a fleet that also has landings data so you no longer have to create two fleets. This led to other refactoring such as "catch" to "landings" and allowing the composition data to be separate from the catch. Now, if you have catch compositions those are based on F, if you have survey compositions those are scaled by q, and you can also have composition data without catch or CPUE and those match the population. Compositions are different for each because survey uses the whole population and landings are just of the retained portion which is not exactly the same due to the impact of natural mortality.
* Fixes the rcpp_fleet json documentation that had out of order bracketing
* Updates estimates tibble snapshot with new reporting names and values

Co-authored-by: Bai Li - NOAA <Bai-Li-NOAA@users.noreply.github.com>
Co-authored-by: Nathan Vaughan <nathanvaughan-NOAA@users.noreply.github.com>
Co-authored-by: Andrea-Havron-NOAA <Andrea-Havron-NOAA@users.noreply.github.com>
Co-authored-by: kellijohnson-NOAA <kelli.johnson@noaa.gov>

Closes #751