feat(logit, inv_logit): Exports logit functions
logit(a, b, x) was already in C++ but now can be accessed from R
through the use of the Rcpp package. This allows for faster computation
of the logit function in R and no duplicated code.

inv_logit(a, b, logit_x) does the reverse transformation of the logit.