chore: Moves collaborative workflow to FIMS
Much of the content was moved to CONTRIBUTING.md but some was also moved
to the README. The documentation vignette was removed because it mainly
contained links to resources that we are trying to consolidate into more
intuitive locations. `check_fims()` was deprecated and the list of items
was moved to CONTRIBUTING.md. Adds information from collaborative
workflow to the README in the tests folder.

Close #947